+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/package.scala
@@ -1,0 +1,5 @@
+package com.twitter.inject.app
+
+package object internal {
+  type ExitFunction = () => Unit
+}