+++ inject/inject-server/src/main/scala/com/twitter/inject/server/PortUtils.scala
@@ -5,36 +5,54 @@ import com.twitter.finagle.ListeningServer
 import com.twitter.finagle.builder.{Server => BuilderServer}
 import java.net.{InetAddress, InetSocketAddress, SocketAddress}
 
+/**
+ * Utilities for formatting Integer ports and SocketAddresses.
+ */
 object PortUtils {
 
+  /** Returns a `host:port` String which is the loopback address and the emphemeral port :0 */
   def ephemeralLoopback: String = {
     loopbackAddressForPort(0)
   }
 
-  def loopbackAddress = {
+  /**
+   * Returns the loopback Inet Address.
+   *
+   * @see [[java.net.InetAddress.getLoopbackAddress]]
+   */
+  def loopbackAddress: String = {
     InetAddress.getLoopbackAddress.getHostAddress
   }
 
-  def loopbackAddressForPort(port: Int) = {
+  /** Returns a `host:port` String for the loopback address and the given port. */
+  def loopbackAddressForPort(port: Int): String = {
     s"$loopbackAddress:$port"
   }
 
+  /** Returns the Integer representation of the bound address for the given [[ListeningServer]] */
   def getPort(server: ListeningServer): Int = {
     getPort(server.boundAddress)
   }
 
+  /** Returns the Integer representation of the given [[SocketAddress]] */
   def getPort(socketAddress: SocketAddress): Int = {
     socketAddress.asInstanceOf[InetSocketAddress].getPort
   }
 
+  /** Returns the Integer representation of the given [[BuilderServer]] */
   def getPort(server: BuilderServer): Int = {
     getSocketAddress(server).asInstanceOf[InetSocketAddress].getPort
   }
 
+  /** Returns the bound address of the given [[BuilderServer]] */
   def getSocketAddress(server: BuilderServer): SocketAddress = {
     server.boundAddress
   }
 
+  /**
+   * Parses the given String as an InetSocketAddress using [[com.twitter.app.Flaggable]]
+   * @see [[com.twitter.app.Flaggable.ofInetSocketAddress]]
+   */
   def parseAddr(addrStr: String): InetSocketAddress = {
     Flaggable.ofInetSocketAddress.parse(addrStr)
   }